refactor rust project
